+++ CHANGELOG.md
@@ -11,6 +11,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - **ThrustController** containing most of the settings from the Utilities window
 - `LaunchToRendezvous()` and `LaunchToTargetPlane()` methods to AscentAutopilot
 - `AutostagingOnce` property to StagingController
+### Fixed
+- `AscentAutopilot#DesiredInclination` property didn't do anything
 
 ## [0.2.0] - 2018-09-07
 ### Added