+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.3.0] - 2018-09-15
 ### Added
 - **DeployableController** for antennas and solar panels
 - **SmartRCS** window
@@ -70,6 +72,7 @@ proximity or reverting a flight to launch
 - **StagingController**
 - **TargetController** to get information about the current target
 
-[Unreleased]: https://github.com/Genhis/KRPC.MechJeb/compare/v0.2.0...HEAD
+[Unreleased]: https://github.com/Genhis/KRPC.MechJeb/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/Genhis/KRPC.MechJeb/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/Genhis/KRPC.MechJeb/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/Genhis/KRPC.MechJeb/commit/6fafaaa41df39a60933d75cfd9c765c5aa8691f7